zac_auto_nippou/pyproject.toml
[project]
name = "zac-auto-nippou"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jpholiday>=1.0.3",
    "playwright>=1.58.0",
    "python-dotenv>=1.0.0",
]
